+++ source/mtree_float_util.h
@@ -7,13 +7,13 @@
 
 #include "mtree_float.h"
 
-float mtree_float_full_distance(mtree_float* first, mtree_float* second);
-float mtree_float_outer_distance(mtree_float* first, mtree_float* second);
+double mtree_float_full_distance(mtree_float* first, mtree_float* second);
+double mtree_float_outer_distance(mtree_float* first, mtree_float* second);
 bool mtree_float_equals(mtree_float* first, mtree_float* second);
 bool mtree_float_overlap_distance(mtree_float* first, mtree_float* second);
 bool mtree_float_contains_distance(mtree_float* first, mtree_float* second);
 bool mtree_float_contained_distance(mtree_float* first, mtree_float* second);
 mtree_float* mtree_float_deep_copy(mtree_float* source);
-float get_float_distance(int size, mtree_float* entries[size], float distances[size][size], int i, int j);
+double get_float_distance(int size, mtree_float* entries[size], double distances[size][size], int i, int j);
 
 #endif
